style: resolve symblic links in `HOMEBREW_CACHE`
On Linux HOMEBREW_CACHE may be set to `$HOME/.cache/Homebrew`. Some
systems, like Fedora Silverblue, have `/home` set up to be a symbolic
link to `/var/home`. While Homebrew generally seems to work fine on
Fedora Silverblue, rubocop causes the `brew style` command to emit
a lot of warnings because of the symbolic link[1]. This can be resolved
by looking up the real path of HOMEBREW_CACHE before passing it to rubocop.

[1] https://github.com/rubocop/rubocop/issues/6228

Co-authored-by: Bo Anderson <mail@boanderson.me>