github_runner_matrix: extend timeout for dependent tests on arm64
We give our ARM runners half the timeout of the Intel runners because
the ARM runners are faster. However, this is no longer true for
dependent testing because we test recursive dependents on ARM but skip
them on Intel.

This means that we can often hit the timeout on ARM but have all jobs
finish on Intel. We can re-run these with the long build label, but
that's a bit wasteful of our limited long build slots.

Instead, let's just use the same timeout value across ARM and Intel
runners when testing dependents: 2 hours.