github_runner_matrix: deploy arm64 Linux runners on PRs with `:arm64_linux` bottles
This will help us avoid losing bottles on version bumps.

Also, let's do the same on Intel macOS 15 and newer. This is essentially
a simpler version of #18363. Instead of trying to use complicated
criteria or a DSL, let's just run the build on Intel macOS 15 if the
formula already has a `:sequoia` bottle.