Add `brew bundle` and `brew services` integration tests
We've had a few regressions in `brew bundle` and `brew services`
recently that would have been caught by actual integration tests.

This is similar to the "default formula" jobs but we want to run outside
of those containers so we can run e.g. systemd.

This will avoid bloating the runtime of already longer-running jobs.